backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
openai==1.12.0
httpx==0.26.0
pydantic==2.5.3
python-multipart==0.0.6
beautifulsoup4==4.12.3
requests==2.31.0
aiofiles==23.2.1
